fix: address critical security, performance, and architecture issues
P0 Security:
- Disable docker.sock mount by default (container escape risk), add gVisor/Firecracker guidance
- Add code safety validation in sandbox_manager.py (block os.system, subprocess, eval, network, etc.)
- Add security_opt and read_only to ai-service container
- Fix CORS allowCredentials + wildcard conflict, use configurable origins

P1 Performance:
- Replace PvpService.findAll() OOM risk with findRandomQuestions() DB-level limit
- Add @Scheduled cleanup for expired GameSession (30min TTL, prevents memory leak)
- Replace ConcurrentLinkedQueue.contains() O(N) with ConcurrentHashMap.newKeySet() O(1)

P2 Architecture:
- Add BehaviorEventController to receive frontend behavior data (fixes 404)
- Unify JwtProvider.getUserId() return type with entity Integer (remove Long/Integer mismatch)
- Fix SSE parser: proper state machine with double-newline event delimiters

P3 Robustness:
- Add SLF4J logging to RankingService Redis exception handlers
- Add multi-choice partial credit scoring with wrong-option penalty
- Dynamic college list loading from API instead of hardcoded data

+++ backend/src/main/java/com/platform/controller/AuthController.java
@@ -33,7 +33,7 @@ public class AuthController {
                 new UsernamePasswordAuthenticationToken(request.getUsername(), request.getPassword()));
 
         User user = userService.findByUsername(request.getUsername());
-        String token = jwtProvider.generateToken(user.getUsername(), (long) user.getId(), user.getRole());
+        String token = jwtProvider.generateToken(user.getUsername(), user.getId(), user.getRole());
         return ResponseEntity.ok(new LoginResponse(
                 token, user.getUsername(), user.getRole(), user.getTotalScore(), user.getRank()));
     }
